refactor: enrichment 파사드 leaf 심볼을 enrichment_network로 분리 (P2-A batch 1)
docs/refactoring-plan-2026-07.md §1.4 1단계에 따라 leaf 함수 8개와 상수 8개를
파사드(common/enrichment.py)에서 common/enrichment_network.py로 이동.

- 함수: is_private_url, _clean_meta_description, _is_low_information_fragment,
  _decode_google_news_base64, _extract_og_metadata,
  _extract_via_{readability,bs4_article,paragraphs}
- 상수: _warned_readability_missing, _gnews_url_cache, _USER_AGENT,
  _NOISE_DESC_PATTERNS, _BOILERPLATE_FRAGMENT_PATTERNS, _BROWSER_UA,
  _IMAGE_META_PATTERNS, _EXCLUDE_CLASS_RE

순환 import 방지를 위해 _clean_meta_description·_is_low_information_fragment를
추출기와 함께 이동(critic 정정 §1.1). 파사드는 하위호환 위해 전 심볼을 재-export
(__all__ 불변, 12개). @patch은 호출자(fetch_page_metadata·resolver·rss_fetcher)가
파사드에 남으므로 common.enrichment.* 유지; Correction C만 network 모듈로 retarget
(_warned_readability_missing monkeypatch + logger).

facade 1186→831줄, enrichment_network 스캐폴드→403줄.
게이트: import OK(순환 없음), 715 passed, ruff clean.
독립 코드리뷰 2관점 APPROVE(호출그래프+REPL로 mock 유효성 실증).
